misc(serve): moved dev server flags temporarily to lib folder
