+++ packages/serve/.gitignore
@@ -1,3 +1,2 @@
 *.js.map
 *.js
-!flags.js
