+++ packages/serve/index.ts
@@ -2,7 +2,7 @@ import * as cmdArgs from "command-line-args";
 import * as WebpackCLI from "../../lib/webpack-cli";
 import * as startDevServer from "./startDevServer";
 import * as argsToCamelCase from "./argsToCamelCase";
-import { devServer } from "./flags";
+import { devServer } from "../../lib/utils/dev-server-flags";
 import { core } from "../../lib/utils/cli-flags";
 
 /**
